cleaned main

+++ Includes/Location.hpp
@@ -1,6 +1,7 @@
 #ifndef LOCATION_HPP
 #define LOCATION_HPP
 
+#include <iterator>
 #include <string>
 #include <vector>
 
@@ -19,7 +20,7 @@ class Location {
         std::string                         _root; 
         std::string                         _index; 
         uint8_t                             _autoindex;
-        std::pair<uint32_t, std::string>    _redirect;
+        std::pair<size_t, std::string>      _redirect;
         bool                                _autorizedMethods[NBR_METHODS];
 
     public:
