+++ Sources/Server.cpp
@@ -46,11 +46,11 @@ void Server::setIp(std::string ip) {
     this->_ip = ip;
 }
 
-void Server::setPort(uint32_t port) {
+void Server::setPort(size_t port) {
     this->_port = port;
 }
 
-void Server::setMaxBodySize(uint32_t size) {
+void Server::setMaxBodySize(size_t size) {
     this->_maxBodySize = size;
 }
 
@@ -58,7 +58,7 @@ void Server::setName(std::string name) {
     this->_serverName = name;
 }
 
-void Server::setError(uint32_t errCode, std::string errPath) {
+void Server::setError(size_t errCode, std::string errPath) {
     if (_errors.count(errCode) == 1) {
         throw DuplicateValueError();
     }
@@ -69,11 +69,11 @@ std::string Server::getIp(void) const {
     return (this->_ip);
 }
 
-std::uint32_t Server::getPort(void) const {
+std::size_t Server::getPort(void) const {
     return (this->_port);
 }
 
-std::uint32_t Server::getMaxBodySize(void) const {
+std::size_t Server::getMaxBodySize(void) const {
     return (this->_maxBodySize);
 }
 
@@ -84,7 +84,7 @@ void Server::setLocation(Location &loc) {
 std::string Server::getName(void) const {
     return (this->_serverName);
 }
-std::map<uint32_t, std::string> Server::getErrors(void) const {
+std::map<size_t, std::string> Server::getErrors(void) const {
     return (this->_errors);
 }
 
@@ -100,8 +100,8 @@ std::ostream &operator<<(std::ostream &out, const Server &serv) {
     out << "Port : " << serv.getPort() << std::endl;
     out << "Max Body Size : " << serv.getMaxBodySize() << std::endl;
     
-    std::map<uint32_t, std::string> errors = serv.getErrors();
-    for (std::map<uint32_t, std::string>::iterator it = errors.begin(); 
+    std::map<size_t, std::string> errors = serv.getErrors();
+    for (std::map<size_t, std::string>::iterator it = errors.begin(); 
             it != errors.end(); it++) {
         out << "Error : [" << it->first << "]" << " : " << it->second << std::endl; 
     }
